# core-common-contract-mgmt-sdk/src/main/resources/api-spec/openapi.yml
{
  "openapi": "3.0.1",
  "info": {
    "title": "core-common-contract-mgmt",
    "description": "Contract Management Core Application",
    "contact": {
      "name": "Firefly Software Foundation",
      "email": "dev@getfirefly.io"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://core.getfirefly.io/contracts",
      "description": "Development Environment"
    },
    {
      "url": "/",
      "description": "Local Development Environment"
    }
  ],
...
  "paths": {
    "/api/v1/contracts/{contractId}/terms/{termId}": {
      "get": {
        "tags": [
          "Contract Terms"
        ],
        "summary": "Get contract term by ID",
        "description": "Returns a contract term based on its ID",
        "operationId": "getContractTermById",
        "parameters": [
          {
            "name": "contractId",
            "in": "path",
            "description": "ID of the contract",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "termId",
            "in": "path",
            "description": "ID of the contract term to retrieve",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal server error"
          },
          "404": {
            "description": "Contract term not found"
          },
          "200": {
            "description": "Successfully retrieved contract term",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractTermDynamicDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Contract Terms"
        ],
        "summary": "Update contract term",
        "description": "Updates an existing contract term with the provided information",
        "operationId": "updateContractTerm",
        "parameters": [
          {
            "name": "contractId",
            "in": "path",
            "description": "ID of the contract",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "termId",
            "in": "path",
            "description": "ID of the contract term to update",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContractTermDynamicDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Contract term successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractTermDynamicDTO"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          },
          "404": {
            "description": "Contract term not found"
          },
          "400": {
            "description": "Invalid contract term data provided"
          }
        }
      },
      "delete": {
        "tags": [
          "Contract Terms"
        ],
        "summary": "Delete contract term",
        "description": "Deletes a contract term based on its ID",
        "operationId": "deleteContractTerm",
        "parameters": [
          {
            "name": "contractId",
            "in": "path",
            "description": "ID of the contract",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "termId",
            "in": "path",
            "description": "ID of the contract term to delete",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Contract term successfully deleted"
          },
          "500": {
            "description": "Internal server error"
          },
          "404": {
            "description": "Contract term not found"
          }
        }
      }
    },
...
    "/api/v1/contracts/{contractId}": {
      "get": {
        "tags": [
          "Contracts"
        ],
        "summary": "Get contract by ID",
        "description": "Returns a contract based on its ID",
        "operationId": "getContractById",
        "parameters": [
          {
            "name": "contractId",
            "in": "path",
            "description": "ID of the contract to retrieve",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal server error"
          },
          "200": {
            "description": "Successfully retrieved contract",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractDTO"
                }
              }
            }
          },
          "404": {
            "description": "Contract not found"
          }
        }
      },
      "put": {
        "tags": [
          "Contracts"
        ],
        "summary": "Update contract",
        "description": "Updates an existing contract with the provided information",
        "operationId": "updateContract",
        "parameters": [
          {
            "name": "contractId",
            "in": "path",
            "description": "ID of the contract to update",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContractDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid contract data provided"
          },
          "500": {
            "description": "Internal server error"
          },
          "200": {
            "description": "Contract successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractDTO"
                }
              }
            }
          },
          "404": {
            "description": "Contract not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Contracts"
        ],
        "summary": "Delete contract",
        "description": "Deletes a contract based on its ID",
        "operationId": "deleteContract",
        "parameters": [
          {
            "name": "contractId",
            "in": "path",
            "description": "ID of the contract to delete",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal server error"
          },
          "404": {
            "description": "Contract not found"
          },
          "204": {
            "description": "Contract successfully deleted"
          }
        }
      }
    },
    "/api/v1/contract-term-templates/{termTemplateId}/validation-rules/{validationRuleId}": {
      "get": {
        "tags": [
          "Contract Term Validation Rules"
        ],
        "summary": "Get contract term validation rule by ID",
        "description": "Returns a contract term validation rule based on its ID",
        "operationId": "getContractTermValidationRuleById",
        "parameters": [
          {
            "name": "termTemplateId",
            "in": "path",
            "description": "ID of the term template",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "validationRuleId",
            "in": "path",
            "description": "ID of the contract term validation rule to retrieve",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Contract term validation rule not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "200": {
            "description": "Successfully retrieved contract term validation rule",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractTermValidationRuleDTO"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Contract Term Validation Rules"
        ],
        "summary": "Update contract term validation rule",
        "description": "Updates an existing contract term validation rule with the provided information",
        "operationId": "updateContractTermValidationRule",
        "parameters": [
          {
            "name": "termTemplateId",
            "in": "path",
            "description": "ID of the term template",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "validationRuleId",
            "in": "path",
            "description": "ID of the contract term validation rule to update",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContractTermValidationRuleDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Contract term validation rule not found"
          },
          "400": {
            "description": "Invalid contract term validation rule data provided"
          },
          "500": {
            "description": "Internal server error"
          },
          "200": {
            "description": "Contract term validation rule successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractTermValidationRuleDTO"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Contract Term Validation Rules"
        ],
        "summary": "Delete contract term validation rule",
        "description": "Deletes a contract term validation rule based on its ID",
        "operationId": "deleteContractTermValidationRule",
        "parameters": [
          {
            "name": "termTemplateId",
            "in": "path",
            "description": "ID of the term template",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "validationRuleId",
            "in": "path",
            "description": "ID of the contract term validation rule to delete",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Contract term validation rule not found"
          },
          "500": {
            "description": "Internal server error"
          },
          "204": {
            "description": "Contract term validation rule successfully deleted"
          }
        }
      }
    },
...
    "/api/v1/contracts/{contractId}/terms/filter": {
      "post": {
        "tags": [
          "Contract Terms"
        ],
        "summary": "Filter contract terms",
        "description": "Returns a paginated list of contract terms based on filter criteria",
        "operationId": "filterContractTerms",
        "parameters": [
          {
            "name": "contractId",
            "in": "path",
            "description": "ID of the contract",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilterRequestContractTermDynamicDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid filter criteria provided"
          },
          "500": {
            "description": "Internal server error"
          },
          "200": {
            "description": "Successfully retrieved contract terms",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/contracts/filter": {
      "post": {
        "tags": [
          "Contracts"
        ],
        "summary": "Filter contracts",
        "description": "Returns a paginated list of contracts based on filter criteria",
        "operationId": "filterContracts",
        "parameters": [
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilterRequestContractDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid filter criteria provided"
          },
          "500": {
            "description": "Internal server error"
          },
          "200": {
            "description": "Successfully retrieved contracts",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/contract-term-templates/{termTemplateId}/validation-rules/filter": {
      "post": {
        "tags": [
          "Contract Term Validation Rules"
        ],
        "summary": "Filter contract term validation rules",
        "description": "Returns a paginated list of contract term validation rules based on filter criteria",
        "operationId": "filterContractTermValidationRules",
        "parameters": [
          {
            "name": "termTemplateId",
            "in": "path",
            "description": "ID of the term template",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FilterRequestContractTermValidationRuleDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Invalid filter criteria provided"
          },
          "500": {
            "description": "Internal server error"
          },
          "200": {
            "description": "Successfully retrieved contract term validation rules",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginationResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/contract-term-templates/{termTemplateId}/validation-rules": {
      "post": {
        "tags": [
          "Contract Term Validation Rules"
        ],
        "summary": "Create a new contract term validation rule",
        "description": "Creates a new contract term validation rule with the provided information",
        "operationId": "createContractTermValidationRule",
        "parameters": [
          {
            "name": "termTemplateId",
            "in": "path",
            "description": "ID of the term template",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "X-Idempotency-Key",
            "in": "header",
            "description": "Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContractTermValidationRuleDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Contract term validation rule successfully created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContractTermValidationRuleDTO"
                }
              }
            }
          },
          "400": {
            "description": "Invalid contract term validation rule data provided"
          },
          "500": {
            "description": "Internal server error"
          },
          "404": {
            "description": "Term template not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ContractTermDynamicDTO": {
        "required": [
          "contractId",
          "effectiveDate",
          "termTemplateId"
        ],
        "type": "object",
        "properties": {
          "termId": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "contractId": {
            "type": "string",
            "format": "uuid"
          },
          "termTemplateId": {
            "type": "string",
            "format": "uuid"
          },
          "termValueText": {
            "type": "string"
          },
          "termValueNumeric": {
            "type": "number"
          },
          "termValueJson": {
            "$ref": "#/components/schemas/JsonNode"
          },
          "effectiveDate": {
            "type": "string",
            "format": "date-time"
          },
          "expirationDate": {
            "type": "string",
            "format": "date-time"
          },
          "isActive": {
            "type": "boolean"
          },
          "notes": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        }
      },
      "JsonNode": {
        "type": "object"
      },
...
      "ContractDTO": {
        "required": [
          "contractStatus"
        ],
        "type": "object",
        "properties": {
          "contractId": {
            "type": "string",
            "format": "uuid",
            "readOnly": true
          },
          "contractNumber": {
            "maxLength": 255,
            "minLength": 0,
            "type": "string"
          },
          "contractStatus": {
            "type": "string",
            "enum": [
              "DRAFT",
              "PENDING_APPROVAL",
              "ACTIVE",
              "SUSPENDED",
              "TERMINATED",
              "EXPIRED",
              "CANCELLED"
            ]
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "productCatalogId": {
            "type": "string",
            "format": "uuid"
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "readOnly": true
          }
        }
      },
...
      "FilterRequestContractTermDynamicDTO": {
        "required": [
          "pagination"
        ],
        "type": "object",
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/ContractTermDynamicDTO"
          },
          "rangeFilters": {
            "$ref": "#/components/schemas/RangeFilter"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationRequest"
          },
          "options": {
            "$ref": "#/components/schemas/FilterOptions"
          }
        },
        "description": "Generic filter request that includes both filter criteria and pagination"
      },
...
      "RangeFilter": {
        "type": "object",
        "properties": {
          "ranges": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/RangeObject"
            },
            "description": "Map of field names to their range criteria"
          }
        },
        "description": "Represents range filters for numeric fields"
      },
...
    }
  }
}